+++ cmd/root/root.go
@@ -2,9 +2,9 @@ package root
 
 import (
 	"github.com/cli/cli/v2/pkg/cmdutil"
+	"github.com/majikmate/gh-mmc/cmd/archive"
 	"github.com/majikmate/gh-mmc/cmd/clone"
 	"github.com/majikmate/gh-mmc/cmd/initialize"
-	"github.com/majikmate/gh-mmc/cmd/leave"
 	"github.com/majikmate/gh-mmc/cmd/sync"
 	"github.com/spf13/cobra"
 )
@@ -18,7 +18,7 @@ func NewRootCmd(f *cmdutil.Factory) *cobra.Command {
 	cmd.AddCommand(initialize.NewCmdInit(f))
 	cmd.AddCommand(clone.NewCmdClone(f))
 	cmd.AddCommand(sync.NewCmdSync(f))
-	cmd.AddCommand(leave.NewCmdLeave(f))
+	cmd.AddCommand(archive.NewCmdArchive(f))
 
 	return cmd
 }
